add a description of the mapping for wheel 1353909
Obtained an ASCII Courier 10 DP PC (reorder number 1353909) and checked
the character set. The first column is the printwheel address, the
second column is the printed character, and the third column is the
ASCII code. This printwheel will let me print a full ASCII character
set on the wheelwriter.

The difference between a standard wheelwriter typewheel and this one are
9 glyph substitutions:

0x3a: centmark becomes ^
0x3c: plusminus becomes `
0x42: super-3 becomes \
0x43: super-2 becomes |
0x44: degree becomes ~
0x45: sectionmark becomes <
0x46: paragraphmark becomes >
0x47: 1/2 becomes }
0x48: 1/4 becomes {

Signed-off-by: Russell Senior <russell@personaltelco.net>